+++ ruleengine-core/src/test/kotlin/ruleengine/manifest/ManifestPathResolverTest.kt
@@ -1,4 +1,4 @@
-package ui.editor.rules
+package ruleengine.manifest
 
 import java.nio.file.Path
 import kotlin.test.Test